Removed unused (and untested) code and add a application shutdown hook to dispose the resources
